share validation logic
