+++ static/js/src/advantage/users/components/AddNewUser/AddNewUserForm.tsx
@@ -4,42 +4,23 @@ import {
   ActionButton,
   Input,
   Select,
-  ValueOf,
 } from "@canonical/react-components";
 import { Formik, Form, Field } from "formik";
 
 import { HandleNewUserSubmit, UserRole } from "../../types";
 import { userRoleOptions } from "../../constants";
+import {
+  getErrorMessage,
+  validateEmail,
+  validateRequired,
+  SubmissionErrorMessage,
+} from "../../utils";
 
 interface Values {
   email: string;
   role: UserRole;
   name: string;
 }
-
-const validateRequired = (value: string): string | undefined =>
-  !value ? "This field is required." : undefined;
-
-const validateEmail = (value: string): string | undefined =>
-  validateRequired(value) ||
-  !/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(value)
-    ? "Must be a valid email."
-    : undefined;
-
-const errorMessages = {
-  email_already_exists: "Account with this email address exists.",
-  default: "An unknown error has occurred.",
-} as const;
-
-type SubmissionErrorMessageKey = keyof typeof errorMessages;
-type SubmissionErrorMessage = ValueOf<typeof errorMessages>;
-
-const getErrorMessage = (
-  error: SubmissionErrorMessageKey | string = "default"
-): SubmissionErrorMessage =>
-  Object.keys(errorMessages).includes(error)
-    ? errorMessages[error as SubmissionErrorMessageKey]
-    : errorMessages.default;
 
 export const AddNewUserForm = ({
   handleClose,
